move const higher
